+++ tests/_utilities/__init__.py
@@ -1,0 +1,2 @@
+from .factories import FactoryManager  # noqa
+from .fakes import FakeMeetupAdapter, FakeSlackClient  # noqa
