+++ tests/_utilities/factories/event_details.py
